--- Assets/Scripts/Weapon/Projectile/ProjectileUsageExample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ad71b0e8be7c02b4ba1851b6bc8d112a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
